Add files via upload
This Database created on SQL was a Sales project based on the famous database AdventureWorks2019 from Microsoft.

I decided to create a code and share this project on SQL to let people learn from it.

This database is not as bigger as AdventureWorks2019, it is a database made of 10 different tables linked to each other.
After the step of creation of the tables, I did a step only to add the data into the table. I added the data from the AdventureWorks2019 database.